+++ R/threshold_summary.R
@@ -224,7 +224,8 @@ threshold_summary.swmpr <- function(swmpr_in
                         , year = c(mn_yr:mx_yr)
                         , stringsAsFactors = FALSE)
 
-    dat_grp <- left_join(dummy, summary, by = "dummy")
+    # dat_grp <- left_join(dummy, summary, by = "dummy")
+    dat_grp <- left_join(dummy, summary, by = summary_type)
     dat_grp$count[is.na(dat_grp$count)] <- 0
 
     dat_grp$grp_join <- factor(dat_grp$grp_join)